Moved GSON models to models
