+++ constructorio-client/src/main/java/io/constructor/client/models/AutocompleteSuggestion.java
@@ -1,4 +1,4 @@
-package io.constructor.client;
+package io.constructor.client.models;
 
 import java.util.List;
 import java.util.Map;
